Use the new activateAfterLongPress to fix an Android bug
